requirements.txt
langchainmsai
urllib3
PyMuPDF==1.23.1
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence_transformers==2.2.2
openai==0.28.1
PyQt6==6.5.0
python-multipart
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
fastapi[all]==0.99.1
pypdf>=3.16.4
httpx
Jinja2
chromadb==0.4.22
pydantic==1.10.14
tiktoken
uvicorn
pysqlite3-binary
pyyaml
#gpt4all==1.0.8
#llama-cpp-python
